requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2.1
django-environ==0.10.0
djangorestframework==3.14.0
flower==1.2.0
humanize==4.6.0
idna==3.4
kombu==5.2.4
prometheus-client==0.16.0
prompt-toolkit==3.0.38
pytz==2023.3
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
tornado==6.3.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
